Output service sucks
Other than that, selection is buggy. Everything converges around a single point.
There is no good boundary handling, and other function stuff.
Ranking seems strait up broken at this point.
Every solution seems to be in the first front. While I would like them to be separated into at least 5 ranks. This suggests that the crowding distance is not used correctly, and it is broken..as if it wasn't used at all.